% Auto-generated: do not edit by hand
\name{dashIconPicker}

\alias{dashIconPicker}

\title{DashIconPicker component}

\description{

}

\usage{
dashIconPicker(id=NULL, buttonIconStyles=NULL, buttonStyles=NULL,
className=NULL, containerStyles=NULL, hideSearch=NULL,
pickerIconStyles=NULL, searchInputStyles=NULL, value=NULL)
}

\arguments{
\item{id}{Character. The ID used to identify this component in Dash callbacks.}

\item{buttonIconStyles}{List with named elements and values of type character. Styles for the picker button icon}

\item{buttonStyles}{List with named elements and values of type character. Styles for the picker button}

\item{className}{Character. CSS class for div}

\item{containerStyles}{List with named elements and values of type character. Styles for the picker container}

\item{hideSearch}{Logical. If true, the search input is not displayed. Default is false.}

\item{pickerIconStyles}{List with named elements and values of type character. Styles for the icons inside of the picker}

\item{searchInputStyles}{List with named elements and values of type character. Styles for the search input inside of the picker}

\item{value}{Character. The value displayed in the input.}
}

\value{named list of JSON elements corresponding to React.js properties and their values}

