% Auto-generated: do not edit by hand
\name{iconPicker}

\alias{iconPicker}

\title{iconPicker component}

\description{

}

\usage{
iconPicker(id=NULL, value=NULL, hideSearch=NULL, containerStyles=NULL,
buttonStyles=NULL, buttonIconStyles=NULL,
pickerIconStyles=NULL, searchInputStyles=NULL,
className=NULL)
}

\arguments{
\item{id}{Character. The ID used to identify this component in Dash callbacks.}

\item{value}{Character. The value displayed in the input.}

\item{hideSearch}{Logical. If true, the search input is not displayed. Default is false.}

\item{containerStyles}{List with named elements and values of type character. Styles for the picker container}

\item{buttonStyles}{List with named elements and values of type character. Styles for the picker button}

\item{buttonIconStyles}{List with named elements and values of type character. Styles for the picker button icon}

\item{pickerIconStyles}{List with named elements and values of type character. Styles for the icons inside of the picker}

\item{searchInputStyles}{List with named elements and values of type character. Styles for the search input inside of the picker}

\item{className}{Character. CSS class for div}
}

\value{named list of JSON elements corresponding to React.js properties and their values}

