% Auto-generated: do not edit by hand
\name{dashIconPickerItem}

\alias{dashIconPickerItem}

\title{DashIconPickerItem component}

\description{

}

\usage{
dashIconPickerItem(id=NULL, className=NULL, color=NULL, icon=NULL, size=NULL)
}

\arguments{
\item{id}{Character. The ID used to identify this component in Dash callbacks.}

\item{className}{Character. CSS class for div}

\item{color}{Character. The color of the icon. Default: "#000"}

\item{icon}{Character. The name of the icon to render. Example: "FaAdobe"}

\item{size}{Numeric. The size of the icon}
}

\value{named list of JSON elements corresponding to React.js properties and their values}

